Train dataset ready.
I added the data loader for the DCASe dataset. This was copied from
the GANtestbed and then fixed. So now it is a map style dataset, where
the length is determined by the number of wavs loaded. This is set in
the list_dataset file. So I have to manually select the wavs I want
to work with. Very similar to the Sony style.
But then, in the getitem, I have 3 ways of slicing the audio. If I
use random, then I will get a random slice everytime. With the correct
labels.
So I have added a new InfiniteDataLoader, that can take this dataset,
iterate an epoch, where it samples each wav file once. And repeats forever.
I am not 100% happy with how it has to sample each wav for each epoch, but
for now it is OK for training.

When using full, I can then iterate each wav manually outside the dataset
to do the anaylysis and evaliation.

Some issues:
- DCASE 2022 has files with different file lengths. I have not tested this.
- Some harcoded part in the evaluation test

I also fixed some stuff in plotting and utils.