Fixed dataset with multiple audio durations
I worked on the dataset and it seems to be workign ok for the DCASE
2022. So now, the audios have different lengths, and I am storing
these durations so that I can do the random slicing correctly, covering
the whole length of each file.

I also worked on the valdiation part. It seems to be working ok. At least
I can load full clips with labels, and divide them evenly into chunks.
Even when the durations are different, or trimmed.

I briefly tested the multitrack ACCDOA fucntion, and it kind works.
I am not sure if this will be the case if I change the hop size.