Towawrds evaluation metrics and validation steps.
So I have been workign on adding the validaiton metrics and step.
This is a partial commit, so it is not ready yet.
I have a bit of a mess, because I copied the evaluation metrics files
directly from the 2022 baseline. But then I added the warpper from the
sony internship files. I am NOT sure if this two are fully compatible.
It seems to work somwhoe, I can at least generate the indivual csvs
for each validation wav, and compute some metrics.
Wheteher these metrics are right of wrong, I dont know.

Next step I wioll refactator the main code, to get a cleaner code and
more similar to the baseline, in how they call the validaiton step.