Towards more evaluation, and histograms SED
I have continued working on the evaluation metrics. I think it more or
less works now. I have run a few training loops and evalution runs. But
the code in  main is very messy and I want to change it.

I also added code to get the histograms for sound event classes per frames
in the dataset. This includes the dcase2021, dcase2022, and dcase2022_sim
datatsets. There is quite a bit dataset imbalance.

IMORTANT: The valudation is not ok when using overlap. As I found a mistake
for file 42 or 43 of dcase2022 when overlap = 0.5 . I am not sure why.

Next I will refactor the main code a bit.