Baseline mostly done, still a bit messy
I refactored some of the code in main. It is still very messy and
horrible, but less ugly than before. Now it seems to be working ok
for training and validaiton. Still no test set evaluation.
I will maybe try to refactor a bi tmore, I really want to simplify
the iteration of the dataset.

Multi-accdoa is not working.

It also appears to be quite slow, I am not sure why.