Evaluation debugged, mostly. STFT + IV features
So I spent a lot of tiem debuggign the evaliation code. It seems to be
working ok now, at least I am able to run the overfitting code to a single
wav and get more or less ok metrics. Except for the localizaiton part that
is still very bad. But I dont know if there is an eeror in the metric,
or if the prediction is just that bad. I need to debug it.

The ouput csv looks ok.

The multi track ACCDDOA still does not work.

It is still super slow.

I also added features to compute the magnitude STFT and the foa intensity
vectors. This looks more or less ok.