MASSIVE Debugging, evaluation, scheduöer, training
So I did a lot of debugging in the main code. It seems to me that the
overfitting to a single wav now works very well. I can basically memorize
the file, but because it does not has all classes, then the metrics are
not super cool. But if we look at the metrics per class, then it is
very good.

Some of the changes include:

- Verified tha the dataset is loading the correct labels, at least
they look good without multitrack.
- Logging of gradients norm.
- Scheduler is now updated correctly. But I am not super sure about
the value of the learning rates.
- Verified the number of classes for the dcase2021 dataset. Ok.
- Output the class wise values, at the end only.
- Veried macro vs micro average.