DCASE2022 submission, CRNN + some audio augmentation
This commit includes my submission to DCASE2022. In the end I did not
have enough time do not even half the stuff I wanted to try. But it is
done. The systems is the basic CRNN10 from sony, but I am computing the
features manually on the fly, in gpu, super fast. And this allows me
to use a lof of audio domain augmentations. I use mostly spatial mixup
as torch audiomentations for this. I also copied the mixup from our
DCASE 2021 submission.

The code is still very messy.

But now there is evaluation, and it is super fast.

There are many things to come next.