Added support for mACCDOA, but still a bit wrong
I added support for the mACCDOA labels by adapting the code from
the official baseline to my dataloader. So now there is a backend
parameter that determines how the labels are extracted from the CSVs.
I was able to run a full training and validation loop, but the results
do not look very good. I am not sure what is wrong.

It could be my CRNN model.

The evaluation is not supported yet.