SampleCNN ok, dcase resubmission
I fixed most of the problems with the SampleCNN. I have two versions
now, one with the conformer, and one with the GRU as the last part.
I ran a full training with the GRU version, and it seems to match the
baseline, barely.

I also had to resubmit the dcase results, because the csvs need to be
in a different format this year.