Fixing torch-audiomentations
So I found a mistake where I was not using the torch-audiomentations
during the trianign loop. Instead, I was calling spatial mixup twice.
Now that I changed that, there are some issues with the augmentaions.
It seems that the pip package is not updated to the last version of
the code in git, and there are some bug fixes that are not included.
So I will commit my stuff as it is now, and change the code to use
torch-audiomentaitons as a git module. I will fork this so that I
can add my own stuff if needed later.

I also added the notebook that I am using to debug the agumentations.