Fixed torch audiomentations, added p_comp scheduler
So i fixed the torch audiomentations. Now its working as a submodule
And I doid a lot of tests, and it seems to work ok now. I also added
the p_comp as curriculum training.
Using audiomentations is a bit slower, but not terrible.
I am not sure why, because the gpu is not at 100%.

I removed the pitchshift because it is slow, and was not working well
with batch mode. But I am using example mode anyway.