New parameters configs loader and WANDB support
I refactored how the parameters are loaded. So now there is a config
file in /configs, that is a yaml file that has the values for all
parameters. These are default values. And then in the code, the parser
reads the file, but the parameters can be overidden via command line.
So I can fo exmple change the --num_wokers 2 and so.
I am quite happy with this now. But sometimes some parameters are tricky
to debug, becuase the files are not loaded correctly.

Currently, I think the best practice is to write the choices , type, and
help in the parameters.py file. Just like when using a regular parser.
And have the yaml mostly clean.

Also, I am now using WANDB to log stuff, and not just tensorboard.
This is still a bit rough, but it seems to work ok.
But I need to do some more profiling.