Profiling with wandb, first batch of full experiments
Overall, a few small changes. Added a couple of parameters, and prepared
the scripts to run full experiments.
I also spent some time profiling the runs now with the wandb stuff.
It seems that adding wandb does NOT slow the run at all. At least
this was the case when runnin in vrgpu. So am I happy now with:

num_worker = 0
pin_memory = False
wandb = True

Added a parameter for the lr_scheduler_step

Added a paramter for groups, this is useful for wandb and to organize
the runs a bit better.