Added soundfield rotations, debugged dataset pads
I added support for the soundfield rotations. So this is done very
similar to the spatial mixup code, where we rotate the grid of spherical
harmonics. So the rotation is done is speherical domain. The same rotation
is applied to the labels. I added a notebook to test this and it looks
good. Also, it seems to not add too much overhead, as some preliminary
tests show about the same performance as not using them.

I also debuged the dataset, where the chunk size was computed incorrectly
due to rounding errors. Now it seems much more conistent, and it works
for 1.27, 2.55, and 5.11 seconds.