Towards overlapped evaluation
I did some test to do the overalpped evaluation. It is getting there
but its still not working. There are padding issues , it only works
with overlap = 1, or overlap = 0.5. And then the actual evaluation
ios wrong, even when using the olverap = 0.5 . So I am not sure.
Well, its not super wrong, but it is worse than without using it.
So maybe there is a problem when computing the average.