Fixed rotations, and analysis of azi and ele
I debugged and fixed the rotations. There was a problem where the
labes were rotated to the opisite direction of the signal. So I
fixed this by tranposing the rotation matrix. Now it looks good.
I added a small analysis to the notebook to verify this.

I also inluced code to do analysis of the bivariate distributions
of azimuth and elevation. This is interesting as it shows how much
coverage there is in each dataset. Dcase2022 has very sparse coverage
because it has real scenes.

NOTE: the rotations are now comnfigure to do rotations on azimuth only.
Lets see how well this works.