New mel features, and better overlap evaluation
So I worked on two main things here. First, there are now new features,
using mel spectrograms and either interchannel phase difference of the
intensity vector. It seems that mel+IV are quite good. Mel+phase not
much. I also refacored the code, so it is better to not do backprop
through the augmentations and features extracrtion.

Then I fixed the overlapped evalation a bit. It is still not perfect
but better than before. It works ok until the overl is 1/32 when the
chunking of audio and labels do not match.
However there is a problem when using data, I need to DEBUG this.

Some highlights:
---- spatial_mixup
- Removed the old code where I was exploring how to plot the bivariate distributions. The plot code is no inside plots.py.

---- anlaysis
- Tiny change, changed the name of the features after refactoring.

---- feature
- Added code for mel+iv, and mel+phasediff features.
- Also cleaned a few things tahta re not needed anymore.

---- main
- Added support for the new features.
- Added support for the parameter for overlap. This is still not super ready.

---- notebeook_augmentation
- Added the part with the rotations. Now it looks good.