Changed torchsummary to torchinfo
I updated the torchsummary library to the new version which is called
torchinfo. It is basically the same thing, but newer, and with more
features. So it works ok for RNN layers too.

I was also reporting the model sizes incorrectly, as it was not taking
the RNN layers into account, the correct sizes are aorund:

CRNN10 - 7.08 million params. There is no change if we increase the
input frames, these only changes the output size and actiations, but not
the number of params. There is also no change if we change the number of
freq bins, because we are doing frequency pooling before the GRU layers.

SampleCNN_gru - 2.49 M params

SampleCNN (with conformer) - 2.12 M params.