Added curriculum schedulers, prepared dcase paper
I added the curriculum schedulers including: fixed, linear, loss and
seld metric. I also added the p_comp parameter to the rotations and
spatial mixup, so that it can applied stochasticly, the same as the
audio domain augmentations. And this p_comp parameter is controlled
with the curriculum scheduler.

I also fixed a bug in the features, where the mel spec sometimes where
giving me some inf values, possibly when the db where computed for empty
frames. This was generating NaNs in the validation loss.

I am now running the expeirments for the workshiop paper.