Added best metrics to wandb
I added some metrics to record the best values and iteration for
the run. This is so that I can easily find the best results of each
run, when reading the table in wandb. This is also very useful when
export to csv, that I can load to an excel and format for the paper.