PARTIAL COMMIT - Added baseline backed for dataset
I have been debugging the dataset, now I added the baseline backed
so that the labels are loaded very similarly to the baseline code
for singel ACCDOA.
This turns to be better than the sony bakcup. However, it does not work
for all cases, for example, it breaks when using inputs of 5.11 secodns.
It also does not work that well for sample cnns.
So the larger the input, the worse it gets.

My intution is that there is soemthing very wrond in the resampling,¨

So Now I will implemente my own backend.