Macro and micro metrics, debugging baseline labels
I added support for both macro and micro metrics, so that it is easier to
analyze experiments. Do note that the ER metrics are the same in all
cases.
I have also been debugging the baseline labels, as looking at the
full files looks ok, but the models are not training. I think there
might be a mistake in the evaliation. Possibly there is some class sawpping
or something similar.