Fixed plots, added plots to validation in main
I fixed a couple of bugs in the analysis plots, mainly things related
to saving the plots to disk.
I also added the analysis plots to the validation mode in main. So now,
when running a validation run, we can look at the plots of all results.

Very useful to evaluate methods.