Fixed plots, and analysis of pre trained runs
I fixed a few things in the plots for the analysis, mainly that the
histograms now are proprotional instead of showing the full count of values
so that it is easier to compared datasets.

I also refactored the code a bit so that it is easier to run the analsysis
for the pretrained models. Today I ran tha analysuis of a few select run
from table03, dcase 2021. Jigsaw is looking good.