Addded file to process results of runs for tables
The file is used to parse the results of a runs and aggregate them
for the tables in the paper. This is still a bit messy, as I have to
run it in debug mode and manually stop and copy the dataframe to
excel, and do the formatting there. But it is useful.