Fixed analysis plots, BUG in evaluation code
I fixed a couple of things in the plots for the analysis. Nothing
major. NOTE: there are some plots disabled for now.

IMPORTANT: I updated the evaluation code with the code from the
sony internship. I dont know which one is the correct one, but now
it seems to get very simialr results to the sony version.
I will use this one fro now.