Towards stratified bootstrapped evaluation
I have been working on a better evaluation to include in the tables
for the paper. The idea is to use stratified bootstrapping, to estimate
the real value of the (mean) metrics for the test dataset.
For this, I am combining checkpoints (from 3 seeds), and wav files into
a pseudo population, and then doing the regular bootstrap there. This is
sampling with replacement, and computing the results.
Finally, I look at the percentiles to get the median and the confidence
intervals.

The problem is that this requires many trials, and it can be quite slow.
It takes about 11 minutes with 100 trials, per system.

So next I will change the code to work in parallel.