Parallel computation of bootstrapped metrics
I have now a working script to compute the bootstrapped metrics of the
results. The script now recevies a chunk id via args, that is used
to find the run in the list_of_runs. Then it computes the bootstrapped
metrics for n_trials. With this chunk_id, we can then run the script
in parallel. The main issue is that this is still limited to a single
run per array job, so we cannot really compute too many trials.

Running for 1000 trials takes about 3, maybe 4 hours.

A better solution would be to parallelize that. But it is a bit tricky.
MMM. maybe add a second chunk_id? To have a run id, and an id for the chunk.
However, we would need to save the intermediate results, to compute the median
and CIs later. So, maybe save the list?

For now this is not needed, as the CIs are tight enough.