Debugging NaNs in DAN training
After running a first run of the DAN training, I was getting some NaNs
in the losses. It seems that the problem was the gradient clipping.
I run a small test with grad clip = 100, instead of 10, and it seems
more estable. Now, I think in the Sony code the grad clip was not
enabled. So maybe it should not be enabled.

I also included a couple of bug fixes here and there, and nicer
plots for the labels.